refactor: move RootSDKViewDimensions$ and Breakpoint$ out of global state
