+++ src/components/Menu/menu.utils.ts
@@ -1,13 +1,14 @@
 import { View } from 'react-native';
-import { RootSDKViewDimensions$ } from '../AppBackground/AppBackground';
 import type { MutableRefObject, RefObject } from 'react';
 import type { Layout } from './menu.types';
 import type { useRootViewRef$ } from '../../hooks/useRootViewRef$';
+import type { useRootSDKViewDimensions$ } from '../../hooks/useRootSDKViewDimensions$';
 
 export function measureViewInWindow(
   viewRef: RefObject<View>,
   layout: MutableRefObject<Layout>,
-  rootViewRef$: ReturnType<typeof useRootViewRef$>
+  rootViewRef$: ReturnType<typeof useRootViewRef$>,
+  rootSDKViewDimensions$: ReturnType<typeof useRootSDKViewDimensions$>
 ) {
   return new Promise<Layout>((resolve, _reject) => {
     // const i = Date.now();
@@ -16,7 +17,7 @@ export function measureViewInWindow(
       return;
     }
     viewRef.current?.measureLayout(rootViewRef, (left, top, width, height) => {
-      const sdkrootview = RootSDKViewDimensions$.peek();
+      const sdkrootview = rootSDKViewDimensions$.peek();
       const _layout = {
         x: left,
         y: top,
