+++ src/components/Menu/menu.windows-paper.tsx
@@ -13,6 +13,7 @@ import { Pressable, StyleSheet, View } from 'react-native';
 import { useCallback, useRef, useState } from 'react';
 import { useRootViewRef$ } from '../../hooks/useRootViewRef$';
 import type { Layout, MenuProps } from './menu.types';
+import { useRootSDKViewDimensions$ } from '../../hooks/useRootSDKViewDimensions$';
 
 export const Menu = function Menu(props: MenuProps) {
   const {
@@ -37,6 +38,7 @@ export const Menu = function Menu(props: MenuProps) {
   });
   const childrenContainerRef = useRef<View>(null);
   const rootViewRef$ = useRootViewRef$();
+  const rootSDKViewDimensions$ = useRootSDKViewDimensions$();
 
   const showOnLeft = () => layout.current.x < layout.current.ww / 2;
   const showOnTop = () => layout.current.y < layout.current.wh / 2;
@@ -47,8 +49,14 @@ export const Menu = function Menu(props: MenuProps) {
   const pointerEvents = enablePointer ? undefined : 'none';
 
   const measureChildrenPosition = useCallback(
-    () => measureViewInWindow(childrenContainerRef, layout, rootViewRef$),
-    [rootViewRef$]
+    () =>
+      measureViewInWindow(
+        childrenContainerRef,
+        layout,
+        rootViewRef$,
+        rootSDKViewDimensions$
+      ),
+    [rootViewRef$, rootSDKViewDimensions$]
   );
 
   const open = () => {
