+++ src/hooks/useRootSDKViewDimensions$.ts
@@ -1,0 +1,6 @@
+import { useContext } from 'react';
+import { AppBackgroundContext } from '../components/AppBackground/AppBackground';
+
+export function useRootSDKViewDimensions$() {
+  return useContext(AppBackgroundContext).rootSDKViewDimensions$;
+}
